+++ Models/Persona.cs
@@ -9,7 +9,7 @@ namespace Sistema_de_gestion_de_empleados.Models
     {
         public string Nombre { get; set; }
         public string Apellido { get; set; }
-        public byte Edad { get; set; }
+        public int Edad { get; set; }
 
         public abstract void MostrarInformacion();
     }
